initial
Adding a login branch where the user sets their school on first login.
Needs to check for the existence of the user in the `users/` ref and then
for a node named `building`. If it exists, login normally. If not, inject
a form to choose a building.

This will build toward an admin/presenter view where those users can search
and filter by building.