(UI) Display courses based on type
On initial load, in person courses are shown to the user, sorted by date
on the frontent. This needs to be done in the back to minimize data transfer.
In person courses only display future events.

Online courses are called using the toggle switch from the user. All courses
are shown, regardless of date.