Trainer course list hotfix!
Only the POC email was shown the training list, but sometimes the
submitter needs to see ths list. This fix checks for a `subimttedBy`
key in the course. If it matches the current user, that course is added
to the dashboard.

This fix polls all courses in the ref, which is a big download. A better fix
will be to poll the current user's `trainer` ref to only pull those courses
listed.

TODO:
- Post `submittedBy` for all current sessions from the master list.
- Update `trainer` ref to include `submittedBy` key
- Only pull listed courses, by date, from the `courseDb` ref based on the
  individual trainer's list.