Feature/trainer page (#9)
Merge `feature/trainer-page` into master. 

Fixes #6.
---
* Update button placement and tooltip activity

* Start framework for cancelling a course

+ Form field logic to show full field key
+ method to destroy a course

The new method requires the user to enter the exact name of the course
in order to delete from the database. This action is non-reversible.

TODO:
- Logic to remove the course from each ref in Firebase
- Automate an email to all participants that the course has been cancelled

* Update button placement and tooltip activity

* Start framework for cancelling a course

+ Form field logic to show full field key
+ method to destroy a course

The new method requires the user to enter the exact name of the course
in order to delete from the database. This action is non-reversible.

TODO:
- Logic to remove the course from each ref in Firebase
- Automate an email to all participants that the course has been cancelled

* Set course `active` key to false for cancelling

* Remove users from a cancelled course

When a course is set to inactive (`false` flag) all registered users of
that course have it removed from `users/{userId}/regs` automatically. This
updates their course dashboard accordingly to show the cancellation.

* Deactivate a course

Automatically send a cancellation notice to all registered users and
remove the course ID from the `users/{userId}/regs` ref. This removes
the course from the logged in view.

* update endpoint

* Complete backend cancellation work

When a course is cancelled, all users have the `courseId` removed
from their list. An automated email is sent to all registered users.
The course is not deleted from the `courses` ref, which allows for
reinstatement of the workshop.

TODO:
- Reinstate method

* Add workshop status to dashboard.